finish implementation (without support for spaces in arguments)
