+++ core/src/implementations/minecraft/server.rs
@@ -24,6 +24,7 @@ use crate::types::Snowflake;
 use crate::util::{dont_spawn_terminal, list_dir};
 
 use super::r#macro::resolve_macro_invocation;
+use super::util::create_java_launch_cmd;
 use super::{Flavour, ForgeBuildVersion, MinecraftInstance};
 use tracing::{error, info, warn};
 
@@ -113,111 +114,22 @@ impl TServer for MinecraftInstance {
                 .join("java")
         };
 
-        let server_start_command = if config.custom_cmd.is_none()
+        let mut server_start_command = if config.custom_cmd.is_none()
             || (config.custom_cmd.is_some() && config.custom_cmd.clone().unwrap().is_empty())
         {
-            let mut server_start_command = Command::new(&jre);
-            let server_start_command = server_start_command
-                .arg(format!("-Xmx{}M", config.max_ram))
-                .arg(format!("-Xms{}M", config.min_ram))
-                .args(
-                    &config
-                        .cmd_args
-                        .iter()
-                        .filter(|s| !s.is_empty())
-                        .collect::<Vec<&String>>(),
-                );
-
-            let server_start_command = match &config.flavour {
-                Flavour::Forge { build_version } => {
-                    let ForgeBuildVersion(build_version) = build_version
-                        .as_ref()
-                        .ok_or_else(|| eyre!("Forge version not found"))?;
-                    let version_parts: Vec<&str> = config.version.split('.').collect();
-                    let major_version: i32 = version_parts[1]
-                        .parse()
-                        .context("Unable to parse major Minecraft version for Forge")?;
-
-                    if 17 <= major_version {
-                        let forge_args = match std::env::consts::OS {
-                            "windows" => "win_args.txt",
-                            _ => "unix_args.txt",
-                        };
-
-                        let mut full_forge_args = std::ffi::OsString::from("@");
-                        full_forge_args.push(
-                            self.path_to_instance
-                                .join("libraries")
-                                .join("net")
-                                .join("minecraftforge")
-                                .join("forge")
-                                .join(build_version.as_str())
-                                .join(forge_args)
-                                .into_os_string()
-                                .as_os_str(),
-                        );
-
-                        server_start_command.arg(full_forge_args)
-                    } else if (7..=16).contains(&major_version) {
-                        let files = list_dir(&self.path_to_instance, Some(false))
-                            .await
-                            .context("Failed to find forge.jar")?;
-                        let forge_jar_name = files
-                            .iter()
-                            .find(|p| {
-                                p.extension().unwrap_or_default() == "jar"
-                                    && p.file_name()
-                                        .unwrap_or_default()
-                                        .to_str()
-                                        .unwrap_or_default()
-                                        .starts_with(format!("forge-{}-", config.version,).as_str())
-                            })
-                            .ok_or_else(|| eyre!("Failed to find forge.jar"))?;
-                        server_start_command
-                            .arg("-jar")
-                            .arg(&self.path_to_instance.join(forge_jar_name))
-                    } else {
-                        // 1.5 doesn't work due to JRE issues
-                        // 1.4 doesn't work since forge doesn't provide an installer
-                        let files = list_dir(&self.path_to_instance, Some(false))
-                            .await
-                            .context("Failed to find minecraftforge.jar")?;
-                        let server_jar_name = files
-                            .iter()
-                            .find(|p| {
-                                p.extension().unwrap_or_default() == "jar"
-                                    && p.file_name()
-                                        .unwrap_or_default()
-                                        .to_str()
-                                        .unwrap_or_default()
-                                        .starts_with("minecraftforge")
-                            })
-                            .ok_or_else(|| eyre!("Failed to find minecraftforge.jar"))?;
-                        server_start_command
-                            .arg("-jar")
-                            .arg(&self.path_to_instance.join(server_jar_name))
-                    }
-                }
-                _ => server_start_command
-                    .arg("-jar")
-                    .arg(&self.path_to_instance.join("server.jar")),
-            };
-
-            let server_start_command = server_start_command.arg("nogui");
-
-            server_start_command
+            create_java_launch_cmd(config.clone(), jre, self.path_to_instance.clone())
+                .await
+                .unwrap()
         } else {
             let mut args: Vec<String> = config
                 .custom_cmd
                 .unwrap()
-                .split(" ")
-                .map(|c| String::from(c))
+                .split(' ')
+                .map(String::from)
                 .collect();
             let mut server_start_command = Command::new(args.remove(0));
-            let server_start_command = server_start_command.args(
-                &config
-                    .cmd_args
-                    .iter()
+            server_start_command.args(
+                args.iter()
                     .filter(|s| !s.is_empty())
                     .collect::<Vec<&String>>(),
             );
